+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Deregister removed layers in interactivity
 - Remove map listeners when there are no layers remaining in interactivity
 - Do not always add CARTO_VL_OTHERS when getting legend data from the buckets expression
+- Fix memory exhausted error using `buckets` with a long list
 
 ## [1.2.6] 2019-05-27